Replaced self-created IntList with the one downloaded from exercise folder
